Fixed death status packet and PacketScreenSize, added PacketTargetedSkill.
- Fixed: The death status packet (sent by the client when the server sent the complementary packet, enabled with the PacketDeathAnimation ini setting) was incorrectly parsed, desynchronizing the input packet queue and making Sphere think it was receiving malformed or unknown packets (Issue #181).
- Fixed: Tooltip info for spawns.
- Fixed: PacketScreenSize (0xBF.0x05) parsing incorrectly the received data.
- Added: Packet UseTargetedSkill (0xBF.0x2E) functionality.
- Added: In debug builds, when an unknown packet is received, its data is dumped to the console and to the log.
- Fixed: EFFECT with id of 4 caused an exception.

Fixed wrong m_iCombatArcheryMovementDelay default value.

+++ src/game/CServerConfig.h
@@ -245,7 +245,7 @@ public:
 	int  m_iArcheryMaxDist;			// Max distance allowed for archery.
 	int  m_iArcheryMinDist;			// Min distance required for archery.
 	int64 m_iHitsUpdateRate;			// how often (seconds) send my hits updates to visible clients..
-	int64 m_iCombatArcheryMovementDelay; // If COMBAT_ARCHERYCANMOVE is not enabled, wait this much tenth of seconds (minimum=0) after the player stopped moving before starting a new attack..
+	int  m_iCombatArcheryMovementDelay; // If COMBAT_ARCHERYCANMOVE is not enabled, wait this much tenth of seconds (minimum=0) after the player stopped moving before starting a new attack..
 	int  m_iCombatDamageEra;		// define damage formula to use on physical combat.
 	int  m_iCombatHitChanceEra;		// define hit chance formula to use on physical combat.
 	int  m_iCombatSpeedEra;			// define swing speed formula to use on physical combat.
